Pass time into in_allowed_window
Before this patch, the current time is calculated inside
in_allowed_window/2, which makes testing it difficult, as the
calculations depend on when tests are run.

This change factors the time as an argument of in_allowed_window, while
the tests continue to test against the current time that they are run.